backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
Django==5.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.2.0
PyJWT==2.12.1
python-dotenv==1.2.2
qrcode==8.2
razorpay==2.0.1
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
